Create 2_5_3_PythonProject:PillowTesseractOpenCV
This is the final project of Coursera course PythonProject: Pillow, Tesseract, and OpenCV. It requires learners to take a zip file of images and process them: the files in the ZIP files are newspaper images, and learners will write python code which allows one to search through the images looking for the occurrences of keywords and faces. For example, if you search for "pizza", it will return a contact sheet of all of the faces which were located on the newspaper page which mentions "pizza".